refactor(db): Move webhook thread reads into dedicated class
* Add DbWebhookReader
* Mark multiple use of `DbWrapper.quests_from_db`
